Split the alert engine out of Lite MainWindow
Extracted the contiguous alert-engine block (CheckPerformanceAlerts, the
SendDetectedAlertAsync sender, the eight Build*Context builders, and their
static helpers — TruncateText/ContextToDetailText/IsDeadlockExcluded/
FormatLowDiskThreshold/FormatDuration) verbatim into a new
Lite/MainWindow.AlertEngine.cs partial. Mirrors the Dashboard MainWindow split.

Same class, same namespace; the alert state fields + the startup watermark-seeding
helper stay in MainWindow.xaml.cs and remain accessible to the partial. Pure
partial-class move — behavior-neutral (main-file diff is deletion-only), no
call-site changes; the live alert path is unchanged.

MainWindow.xaml.cs 2712 -> 1602; MainWindow.AlertEngine.cs 1143. Solution builds
clean; Lite.Tests (545) pass.
